remove impls, fix clippy warnings, move test to new location
